api/requirements.txt
fastapi
uvicorn
psycopg2-binary
requests
protobuf
gtfs-realtime-bindings
